feat(order): surface order detail overlay from status page
